Rename ImpactWidget to InterventionWidget
